LayoutTests/platform/ios-sim-deprecated/accessibility/element-paths-expected.txt

 
This tests SVG group elements are accessible and that the svg:title element is returned properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


SVG path description
Start Path
	Move to point
	Line to
	Line to
	Close

Map1 path description
Start Path
	Move to point
	Line to
	Line to
	Line to
	Line to
	Close

Map2 path description
Start Path
	Move to point
	Curve to
	Curve to
	Curve to
	Curve to
	Close

PASS successfullyParsed is true

TEST COMPLETE
